clean up and modularize gas reactions and effects
